implement furthure authorization and make Swagger2Markup work

+++ src/main/java/com/professionalbugcoding/unicodesc/exception/CommonJsonException.java
@@ -3,24 +3,24 @@ package com.professionalbugcoding.unicodesc.exception;
 public class CommonJsonException extends RuntimeException {
     private static final long serialVersionUID = 2L;
     // 异常编码
-    private Long code;
+    private Integer code;
     // 异常自定义信息
     private String customMsg;
 
     public CommonJsonException() {
     }
 
-    public CommonJsonException(Long code, String customMsg) {
+    public CommonJsonException(Integer code, String customMsg) {
         super();
         this.code = code;
         this.customMsg = customMsg;
     }
 
-    public Long getCode() {
+    public Integer getCode() {
         return code;
     }
 
-    public void setCode(Long code) {
+    public void setCode(Integer code) {
         this.code = code;
     }
